requirements.txt
pytest==8.4.1
playwright==1.55.0
python-dotenv==1.1.1
pytest-html==4.1.1